requirements.txt
attrs==16.2.0
blessings==1.6
bpython==0.15
cffi==1.8.3
click==6.6
cryptography==1.5
cssselect==0.9.2
curtsies==0.2.9
Flask==0.11.1
Flask-Login==0.3.2
Flask-PyMongo==0.4.1
google-api-python-client==1.5.4
greenlet==0.4.10
httplib2==0.9.2
idna==2.1
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
oauth2client==4.0.0
parsel==1.0.3
Pillow==3.3.1
pkg-resources==0.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
Pygments==2.1.3
pymongo==2.8
pyOpenSSL==16.1.0
PyYAML==3.12
queuelib==1.4.2
requests==2.11.1
retrying==1.3.3
rsa==3.4.2
Scrapy==1.1.2
service-identity==16.0.0
six==1.10.0
Twisted==16.4.1
uritemplate==3.0.0
urllib3==1.18
uWSGI==2.0.14
w3lib==1.15.0
wcwidth==0.1.7
Werkzeug==0.11.11
zope.interface==4.3.2
